% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapIdxs.R
\name{BetaIdxs}
\alias{BetaIdxs}
\title{BetaIdxs finds the indices from the overlap data for the beta region}
\usage{
BetaIdxs(OverlapData)
}
\arguments{
\item{OverlapData}{PlaceHolder}

\item{AlpBet}{PlaceHolder}
}
\value{
PlaceHolder
}
\description{
BetaIdxs finds the indices from the overlap data for the beta region
}
\examples{
#
}
