% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxplotAbundances.R
\name{BoxplotAbundances}
\alias{BoxplotAbundances}
\title{BoxplotAbundances produces a paired ggplot of glycopeptide abundances}
\usage{
BoxplotAbundances(
  filename1,
  filename2,
  normvector = list("None", "None"),
  group1,
  group2,
  PlotTitle,
  axislabelsize = 6,
  zerofill = FALSE
)
}
\arguments{
\item{filename1}{The First .csv matrix RAMZIS_formatted file to be entered}

\item{filename2}{The second .csv matrix RAMZIS_formatted file to be entered}

\item{normvector}{The normalization vectors of the files in list format. Default=list("None","None")}

\item{group1}{Name of the first sample group. Corresponds to first file. Use string}

\item{group2}{Name of the second sample group. Corresponds to second file. Use string}

\item{PlotTitle}{Title of the plot}

\item{axislabelsize}{size of axis text}

\item{zerofill}{If glycopeptides are absent from one of the samples, they will be filled in with zero values. Default=FALSE}
}
\value{
ggplot object
}
\description{
BoxplotAbundances produces a paired ggplot of glycopeptide abundances
}
\examples{
#
}
