% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombWRep.R
\name{CombWRep}
\alias{CombWRep}
\title{CombWRep determines the number of combinations possible with replacement}
\usage{
CombWRep(n, r)
}
\arguments{
\item{n}{number of possible members}

\item{r}{number of sampled points}
}
\value{
number possible
}
\description{
CombWRep determines the number of combinations possible with replacement
}
\examples{
#if there are 8 samples and we want to sample 4 with replacement,
#how many possible combinations are there
CombWRep(8,4)
}
