% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContributionCheckComp.R
\name{ContributionCheckComp}
\alias{ContributionCheckComp}
\title{ContributionCheckComp part of the QC set of functions}
\usage{
ContributionCheckComp(SimObj, ExclusionList = "Default")
}
\arguments{
\item{SimObj}{similarity object}

\item{ExclusionList}{list of already failed}
}
\value{
More QC data
}
\description{
ContributionCheckComp part of the QC set of functions
}
\examples{
#
}
