% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContributionCheckInternal.R
\name{ContributionCheckInternal}
\alias{ContributionCheckInternal}
\title{ContributionCheckInternal}
\usage{
ContributionCheckInternal(SimObj, IntA, IntB, ks = F)
}
\arguments{
\item{SimObj}{SimilarityObject}

\item{IntA}{Internal Similarity Object of group 1}

\item{IntB}{Internal Similarity Object of group 2}

\item{ks}{Use of KS test to determine separation. Default=FALSE. Highly Permissive}
}
\value{
contributions
}
\description{
ContributionCheckInternal
}
\examples{
#
}
