% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContributionCheckJoint.R
\name{ContributionCheckJoint}
\alias{ContributionCheckJoint}
\title{ContributionCheckJoint QC Pipeline component}
\usage{
ContributionCheckJoint(SimObj, ExclusionList = "Default", ks = F)
}
\arguments{
\item{SimObj}{Test-Null similarity object}

\item{ExclusionList}{failures}

\item{ks}{KSTest indicator for separation. DEFAULT=FALSE}
}
\value{
QC data
}
\description{
ContributionCheckJoint QC Pipeline component
}
\examples{
#
}
