% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContributionQualityCheckFunction.R
\name{ContributionQualityCheckFunction}
\alias{ContributionQualityCheckFunction}
\title{ContributionQualityCheckFunction}
\usage{
ContributionQualityCheckFunction(
  SimilarityObj,
  InternalSimObj1,
  InternalSimObj2
)
}
\arguments{
\item{SimilarityObj}{TestnullSimilarity Object}

\item{InternalSimObj1}{Internal Similarity object of first group}

\item{InternalSimObj2}{Internal similarity object of second group}
}
\value{
Contribution QC
}
\description{
ContributionQualityCheckFunction
}
\examples{
#
}
