% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContributionQualityCheckFunctionV3.R
\name{ContributionQualityCheckFunctionV3}
\alias{ContributionQualityCheckFunctionV3}
\title{ContributionQualityCheckFunction for use with RankingQualityFunctionV3}
\usage{
ContributionQualityCheckFunctionV3(
  TestSimObj,
  InternalSimObj1,
  InternalSimObj2,
  QualityInfo = "WeightedContributions"
)
}
\arguments{
\item{TestSimObj}{Test Similarity Object}

\item{InternalSimObj1}{Internal Similarity object of first group}

\item{InternalSimObj2}{Internal similarity object of second group}

\item{QualityInfo}{Default="WeightedContributions" Alternative is "Numerator"}
}
\value{
Contribution QC
}
\description{
ContributionQualityCheckFunction for use with RankingQualityFunctionV3
}
\examples{
#
}
