% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContributionQualityCheckSubFunction.R
\name{ContributionQualityCheckSubFunction}
\alias{ContributionQualityCheckSubFunction}
\title{ContributionQualityCheckSubFunction used by CQCFunction}
\usage{
ContributionQualityCheckSubFunction(InternalCont, TestCont)
}
\arguments{
\item{InternalCont}{Internal Ranking info}

\item{TestCont}{Test Ranking info}
}
\value{
QC results
}
\description{
ContributionQualityCheckSubFunction used by CQCFunction
}
\examples{
#
}
