% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContributionQualityCheckSubFunctionV3.R
\name{ContributionQualityCheckSubFunctionV3}
\alias{ContributionQualityCheckSubFunctionV3}
\title{ContributionQualityCheckSubFunction used by CQCFunction}
\usage{
ContributionQualityCheckSubFunctionV3(ReferenceCont, TestCont)
}
\arguments{
\item{ReferenceCont}{Internal Ranking info}

\item{TestCont}{Test Ranking info}
}
\value{
QC results
}
\description{
ContributionQualityCheckSubFunction used by CQCFunction
}
\examples{
#
}
