% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENSITY_RD.R
\name{DENSITY_RD}
\alias{DENSITY_RD}
\title{DENSITY_RD
This is the default density function of RAMZIS and is really not meant for user usage, but is available for validation.}
\usage{
DENSITY_RD(data, lb = -0.1, ub = 1.1, k = 100)
}
\arguments{
\item{data}{A vector of data to be turned into a density function}

\item{lb}{The lower bound of the density function. Default=-0.1}

\item{ub}{The upper bound of the density function. Default=1.1}

\item{k}{The scaling factor by which densitys are adjusted. The deafult converts from proportion to percentage. Default=100}
}
\value{
Density Object akin to base::density()
}
\description{
DENSITY_RD
This is the default density function of RAMZIS and is really not meant for user usage, but is available for validation.
}
\examples{
dist<-rbeta(1000,2,2)
dens<-DENSITY_RD(dist)
}
