% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATHERER_HelperTIDY.R
\name{GATHERER_HelperTIDY}
\alias{GATHERER_HelperTIDY}
\title{GATHERER_HelperTIDY
Get all the information for similarity comparisons}
\usage{
GATHERER_HelperTIDY(gl, combo, df, MVCorrection)
}
\arguments{
\item{gl}{Glycopeptide List Deprecated}

\item{combo}{Sampling}

\item{df}{Dataset}

\item{MVCorrection}{If False, it will ignore missing values rather than count towards the overall average. Default=TRUE}
}
\value{
Gets data for use in matrix similarity comparison
}
\description{
GATHERER_HelperTIDY
Get all the information for similarity comparisons
}
\examples{
#
}
