% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGFormatterFromSym.R
\name{GGFormatterFromSym}
\alias{GGFormatterFromSym}
\title{GGFormatterFromSym takes similarity distributions and turns them in TIDY compatible density data}
\usage{
GGFormatterFromSym(TestDist, RefDist, RefName)
}
\arguments{
\item{TestDist}{Tanimoto similarity vector from the Test Comparison}

\item{RefDist}{Tanimoto similarity vector from the Reference Comparison}
}
\value{
TIDY formatted data
}
\description{
GGFormatterFromSym takes similarity distributions and turns them in TIDY compatible density data
}
\examples{
#
}
