% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGFormatterFromSym.R
\name{GGFormatterGeneral}
\alias{GGFormatterGeneral}
\title{GGFormatterGeneral takes similarity distributions and turns them in TIDY compatible density data}
\usage{
GGFormatterGeneral(TestDist, RefDist)
}
\arguments{
\item{TestDist}{Tanimoto similarity vector from the Test Comparison}

\item{RefDist}{Tanimoto similarity vector from the Null Comparison}
}
\value{
TIDY formatted data
}
\description{
GGFormatterGeneral takes similarity distributions and turns them in TIDY compatible density data
}
\examples{
#
}
