% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGFormatterFromSym.R
\name{GGFormatterInternal}
\alias{GGFormatterInternal}
\title{GGFormatterInternal takes similarity distributions and turns them in TIDY compatible density data}
\usage{
GGFormatterInternal(TestDist, RefDist, IntGrpNum)
}
\arguments{
\item{TestDist}{Tanimoto similarity vector from the Test Comparison}

\item{RefDist}{Tanimoto similarity vector from the Internal Comparison}

\item{IntGrpNum}{A numeric (1 or 2) indicating which file the internal comparison is in relation to}
}
\value{
TIDY formatted data
}
\description{
GGFormatterInternal takes similarity distributions and turns them in TIDY compatible density data
}
\examples{
#
}
