% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlycanSimplifier.R
\name{GlycanSimplifier}
\alias{GlycanSimplifier}
\title{GlycanSimplifier reduces GlycReSoft glycan notation into abbreviated form}
\usage{
GlycanSimplifier(glypep, glysplit = "\\\\{")
}
\arguments{
\item{glypep}{The gylcopeptide to be shortened}
}
\value{
The abbreviated glycopeptide in format PEPTIDE\link{Hex;HexNAc;Fuc;Neu5Ac;Sulfation}
}
\description{
GlycanSimplifier reduces GlycReSoft glycan notation into abbreviated form
}
\examples{
GlycanSimplifier('AMIN(N-Glycosylation)OACIDS{Hex:7; HexNAc:6; Neu5Ac:3}')
}
