% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlycanSplitter.R
\name{GlycanSplitter}
\alias{GlycanSplitter}
\title{GlycanSplitter takes a GlycReSoft formatted glycopeptide and produces a list of its components}
\usage{
GlycanSplitter(glypep)
}
\arguments{
\item{glypep}{Glycopeptide to be split into list}
}
\value{
list with format "GP" is input, "GP_simp" is the simplified form, "pep" is the peptide backbone. Remaining entries are all glycan compositions: Hex, HexNAc, Fuc, Neu5Ac, Sulfate
}
\description{
GlycanSplitter takes a GlycReSoft formatted glycopeptide and produces a list of its components
}
\examples{
#
}
