% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlyCosmosSearchFormatter.R
\name{GlyCosmosSearchFormatter}
\alias{GlyCosmosSearchFormatter}
\title{GlyCosomosSearchFormatter takes a vector of of GlycReSoft formatted glycopeptides and turns it into a .tsv for input into glycosmos composition search engine at https://glycosmos.org/glycans/composition}
\usage{
GlyCosmosSearchFormatter(glycopeptides, outfile, verbose = F)
}
\arguments{
\item{glycopeptides}{vector of glycopeptides}

\item{outfile}{name and location of file to be written to. adds .tsv to given input}

\item{verbose}{Default False. Set to true to get table in console}
}
\value{
.tsv file for glycosmos composition search
}
\description{
GlyCosomosSearchFormatter takes a vector of of GlycReSoft formatted glycopeptides and turns it into a .tsv for input into glycosmos composition search engine at https://glycosmos.org/glycans/composition
}
\examples{
#
}
