% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoading.R
\name{GlycReRead}
\alias{GlycReRead}
\title{Turn multiple GlycReSoft files into a single abundance matrix}
\usage{
GlycReRead(Filelist, Outfilename = NULL, verbose = T)
}
\arguments{
\item{Filelist}{A txt list of CSV files produced by GlycReSoft glycopeptide searches}

\item{Outfilename}{The destination CSV file for the output of GlycReRead: a matrix of abundances}

\item{verbose}{A boolean determining if the user would like the matrix to be returned to the console (T) or not (F). Default is FALSE.}
}
\value{
A matrix containing the raw abundance values from a set of GlycReSoft searches
}
\description{
This takes a filelist and converts it into a single abundance matrix for use with RAMZIS
}
\examples{
#abun_matrix<-GlycReRead(c('sample1.csv','sample2.csv','sample3.csv'),'SampleGroupAbundances.csv')
}
