% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalConfidenceScore.R
\name{InternalConfidenceScore}
\alias{InternalConfidenceScore}
\title{InternalConfidenceScore
Finds the Confidence Score of a Test Similarity Distribution relative to one of its Internal Similarity Distribution}
\usage{
InternalConfidenceScore(ReferenceDis, TestDis, OverlapData = NULL)
}
\arguments{
\item{ReferenceDis}{One of the Internal Similarity Distributions}

\item{TestDis}{The Test Similarity Distribution}

\item{OverlapData}{The known overlap of these distributions from OverlapCalculator()}
}
\value{

}
\description{
InternalConfidenceScore
Finds the Confidence Score of a Test Similarity Distribution relative to one of its Internal Similarity Distribution
}
