% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMembershipProportion.R
\name{InternalMembershipProportion}
\alias{InternalMembershipProportion}
\title{InternalMembershipProportion is for determining the membership of samples to peaks to the Internal similarity distribution}
\usage{
InternalMembershipProportion(IntDist, BootDis1, ModalityList)
}
\arguments{
\item{IntDist}{Internal Similarity Distribution}

\item{BootDis1}{Sampling}

\item{ModalityList}{Output from Modality()}
}
\value{
Sample Proportionship
}
\description{
InternalMembershipProportion is for determining the membership of samples to peaks to the Internal similarity distribution
}
\examples{
#
}
