% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalQuality.R
\name{InternalQuality}
\alias{InternalQuality}
\title{InternalQuality determines quality of Internal similarity}
\usage{
InternalQuality(
  filename,
  BootSet,
  SimilarityObj,
  PlotTitle,
  GroupName,
  Int = NULL,
  kmin = 2,
  rel = TRUE,
  MVCorrection = TRUE,
  mn = FALSE,
  verbose = FALSE,
  legendpos = "topleft",
  legendval = FALSE,
  legendcexval = 1.1,
  xbound = c(0, 1),
  ybound = T,
  logval = T
)
}
\arguments{
\item{filename}{abundance data of sample group}

\item{BootSet}{sampling of sample group}

\item{SimilarityObj}{Similarity Object from Test-Null Comparison}

\item{PlotTitle}{Title for Plot}

\item{GroupName}{Sample Group name}

\item{Int}{Internal Similarity object if already calculated. Default=NULL}

\item{kmin}{minimum number of observations needed to be considered real. Default=1}

\item{rel}{Boolean indicating relative abundance transformation}

\item{MVCorrection}{Boolean missing value correction, if disabled it turns 0s to NAs. Default=TRUE}

\item{mn}{Default=FALSE. In default settings it adjusts the distance scaling by average presence. Setting it to a numeric value will use that as a constant instead. (Recommended 1:2)}

\item{verbose}{Boolean indicating output. Default=FALSE. If TRUE sends Internal Similarity object to console}

\item{legendpos}{Legend position}

\item{legendval}{Legend Value BOOLEAN to produce Original comparison. Default=TRUE}

\item{legendcexval}{Legend text size}

\item{xbound}{Xbound vector default=c(0,1)}

\item{ybound}{ybound vector default has max at max peak height}

\item{logval}{log transform boolean DEFAULT=TRUE}

\item{logopt}{Boolean indicating the use of the log transform before relative scaling of abundance. Default=TRUE}
}
\value{
Plots Internal Similarity
}
\description{
InternalQuality determines quality of Internal similarity
}
\examples{
#
}
