% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalSimilarity.R
\name{InternalSimilarity}
\alias{InternalSimilarity}
\title{InternalSimilarity calculates the similarity of an sample group against itself using bootstraps derived}
\usage{
InternalSimilarity(
  filename,
  BootSet,
  kmin = 1,
  rel = "Within",
  MVCorrection = TRUE,
  mn = FALSE,
  logopt = TRUE,
  normvec = list("None"),
  rel_force = FALSE
)
}
\arguments{
\item{filename}{abundance dataset of sample group}

\item{BootSet}{Sampling of sample group used in Test comparisons}

\item{kmin}{minimum number of observations needed to be considered real. Default=1}

\item{rel}{Boolean indicating relative abundance transformation}

\item{MVCorrection}{Boolean missing value correction, if disabled it turns 0s to NAs. Default=TRUE}

\item{mn}{Default=FALSE. In default settings it adjusts the distance scaling by average presence. Setting it to a numeric value will use that as a constant instead. (Recommended 1:2)}

\item{logopt}{Boolean indicating the use of the log transform before relative scaling of abundance. Default=TRUE}

\item{normvec}{Optional: the normalization vectors to adjust signal abundance between samples. Default=list('None')}

\item{rel_force}{Default=FALSE}
}
\value{
Internal Similarity Object
}
\description{
InternalSimilarity calculates the similarity of an sample group against itself using bootstraps derived
}
\examples{
#
}
