% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalSimilarityFunction.R
\name{InternalSimilarityFunction}
\alias{InternalSimilarityFunction}
\title{InternalSimilarityFunction
For calculating the similarity between a sampling of datasets and itself.
It does not allow chance identical comparisons. (eg. Sampling 1,1,1 will never be compared to Sampling 1,1,1)}
\usage{
InternalSimilarityFunction(
  filename,
  BootSet,
  kmin = 1,
  rel = TRUE,
  MVCorrection = TRUE,
  mn = FALSE,
  logopt = FALSE,
  normvec = c("None")
)
}
\arguments{
\item{filename}{The source file or dataset to be compared}

\item{BootSet}{The samplings to be used}

\item{kmin}{The minimum number of observations needed to be considered real. Default is 1}

\item{rel}{For use in SimDataClean(). Default=TRUE}

\item{MVCorrection}{If False, it will ignore missing values rather than count towards the overall average. Default=TRUE}

\item{mn}{The scaling factor. Default will use 1+mean(presence) of a glycopeptide/identification. Setting this to a number will override that process.}

\item{logopt}{For use in SimDataClean(). Default=FALSE}

\item{normvec}{For use in SimDataClean() Default=c("None")}
}
\value{
Distribution of Internal similarities and their contributions
}
\description{
InternalSimilarityFunction
For calculating the similarity between a sampling of datasets and itself.
It does not allow chance identical comparisons. (eg. Sampling 1,1,1 will never be compared to Sampling 1,1,1)
}
\examples{
#
}
