% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_OUTPUT_GATHERER.R
\name{MAIN_OUTPUT_GATHERER}
\alias{MAIN_OUTPUT_GATHERER}
\title{MAIN_OUTPUT_GATHERER is used by RAMZISMain() to collect and organize all outputs of the similarity comparisons}
\usage{
MAIN_OUTPUT_GATHERER(
  Internal1,
  Internal2,
  Test,
  Null,
  Observed,
  ITO1,
  ITO2,
  NTO,
  Boot1,
  Boot2,
  NullBoot1,
  NullBoot2
)
}
\arguments{
\item{Internal1}{PlaceHolder}

\item{Internal2}{PlaceHolder}

\item{Test}{PlaceHolder}

\item{Null}{PlaceHolder}

\item{Observed}{PlaceHolder}

\item{ITO1}{PlaceHolder}

\item{ITO2}{PlaceHolder}

\item{NTO}{PlaceHolder}

\item{Boot1}{PlaceHolder}

\item{Boot2}{PlaceHolder}

\item{NullBoot1}{PlaceHolder}

\item{NullBoot2}{PlaceHolder}
}
\value{
PlaceHolder
}
\description{
MAIN_OUTPUT_GATHERER is used by RAMZISMain() to collect and organize all outputs of the similarity comparisons
}
\examples{
#
}
