% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManhattanVectorDistance.R
\name{ManhattanVectorDistance}
\alias{ManhattanVectorDistance}
\title{ManhattanVectorDistance determines the manhattan distance of a vector compared to a dataframe}
\usage{
ManhattanVectorDistance(vector1, dataframe1)
}
\arguments{
\item{vector1}{is a vector of numbers}

\item{dataframe1}{is a matrix/dataframe of numbers}
}
\value{
the distance
}
\description{
ManhattanVectorDistance determines the manhattan distance of a vector compared to a dataframe
}
\examples{
#
}
