% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixMerge_Helper.R
\name{MatrixMerge_Helper}
\alias{MatrixMerge_Helper}
\title{MatrixMerge_Helper}
\usage{
MatrixMerge_Helper(m1, m2)
}
\arguments{
\item{m1}{Matrix 1}

\item{m2}{Matrix 2}
}
\value{
A matrix combined by row names
}
\description{
MatrixMerge_Helper
}
