% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modality.R
\name{Modality}
\alias{Modality}
\title{Modality for determining multi-modality}
\usage{
Modality(SimDist, threshold = 0.01)
}
\arguments{
\item{SimDist}{The similarity distribution you would like to find the modality of}

\item{threshold}{The ratio needed to identify a different peak. Default =0.01}
}
\value{
Returns a list of data on the peaks in the similarity distribution
}
\description{
Modality for determining multi-modality
}
\examples{
#
}
