% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalityInterpreter.R
\name{ModalityInterpreter}
\alias{ModalityInterpreter}
\title{ModalityInterpreter}
\usage{
ModalityInterpreter(
  RejectList,
  IntObject1,
  Mode1,
  IntObject2,
  Mode2,
  SimObject = NULL
)
}
\arguments{
\item{RejectList}{Rejection list from modality data}

\item{IntObject1}{Internal similarity object 1}

\item{Mode1}{Modality data of internal 1}

\item{IntObject2}{Internal similarity object 2}

\item{Mode2}{Modality data of internal 2}

\item{SimObject}{Test-Null Similarity Object}
}
\value{
interpretation of modality results
}
\description{
ModalityInterpreter
}
\examples{
#
}
