% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalityTest.R
\name{ModalityTest}
\alias{ModalityTest}
\title{ModalityTest tests for over or under-represented membership in peaks}
\usage{
ModalityTest(SimObject, IntObject1, IntObject2)
}
\arguments{
\item{SimObject}{Test-Null Similarity Object}

\item{IntObject1}{Internal Similarity Object}

\item{IntObject2}{Internal Similarity Object}
}
\value{
Rejection list of samples based on Modality criterion
}
\description{
ModalityTest tests for over or under-represented membership in peaks
}
\examples{
#
}
