% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalityZ.R
\name{ModalityZ}
\alias{ModalityZ}
\title{ModalityZ for determining z scores in modality data}
\usage{
ModalityZ(ModeObject, PropObject, npercent = 0.1)
}
\arguments{
\item{ModeObject}{modality data from Modality()}

\item{PropObject}{proportionality data from a MembershipProportion function}

\item{npercent}{determines allowable variance. (Default is 10\%) Default=0.1}
}
\value{
printed statements about use
}
\description{
ModalityZ for determining z scores in modality data
}
\examples{
#
}
