% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalityZSubset.R
\name{ModalityZSubset}
\alias{ModalityZSubset}
\title{ModalityZSubset is deprecated and doesn't matter}
\usage{
ModalityZSubset(ModeObject, PropObject, npercent = 0.1)
}
\arguments{
\item{ModeObject}{modality object}

\item{PropObject}{proportionality object}

\item{npercent}{threshold}
}
\value{
ignore
}
\description{
ModalityZSubset is deprecated and doesn't matter
}
\examples{
#
}
