% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVCount.R
\name{MVCount}
\alias{MVCount}
\title{MVCount counts the number of missing values}
\usage{
MVCount(vec, val = 0)
}
\arguments{
\item{vec}{Vector to count zeros in}

\item{val}{Value for something to be considered missing equal to. Default=0}
}
\value{
Number of zeros in a vector
}
\description{
MVCount counts the number of missing values
}
\examples{
#
}
