% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization_SubFunction.R
\name{Normalization_SubFunction}
\alias{Normalization_SubFunction}
\title{Normalization_SubFunction produces the normalization of a dataframe}
\usage{
Normalization_SubFunction(df, normvector = "None", logoption = TRUE)
}
\arguments{
\item{df}{Dataframe to be normalized}

\item{normvector}{Vector of NUMERIC values that is equal to sample size/number of df columns. Default='None'}

\item{logoption}{Boolean indicating use of log transformation. Default=TRUE}
}
\value{
Normalized dataframe
}
\description{
Normalization_SubFunction produces the normalization of a dataframe
}
\examples{
#
}
