% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullMembershipProportion.R
\name{NullMembershipProportion}
\alias{NullMembershipProportion}
\title{NullMembershipProportion is for determining the membership of samples to peaks to the Null similarity distribution}
\usage{
NullMembershipProportion(SimDist, BootDis1, BootDis2, ModalityList)
}
\arguments{
\item{SimDist}{Similarity Distribution}

\item{BootDis1}{The first of the null samplings}

\item{BootDis2}{The second of the null samplings}

\item{ModalityList}{The data from Modality about}
}
\value{
The degree of proportion to a peak that each sample contributes to
}
\description{
NullMembershipProportion is for determining the membership of samples to peaks to the Null similarity distribution
}
\examples{
#
}
