% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullSampler_Helper.R
\name{NULLSAMPLER_Helper}
\alias{NULLSAMPLER_Helper}
\title{NullSampler_Helper
Generates samples from two datasets to create null datasets and then removes duplicates.}
\usage{
NULLSAMPLER_Helper(coln1, coln2, sampn = 200)
}
\arguments{
\item{coln1}{number of columns (samples) in first dataset}

\item{coln2}{number of columns (samples) in second dataset}

\item{sampn}{initial number of samplings to begin with. Default=200}
}
\value{
List of dataframes of samplings
}
\description{
NullSampler_Helper
Generates samples from two datasets to create null datasets and then removes duplicates.
}
