% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullSimilarityFunction.R
\name{NullSimilarityFunction}
\alias{NullSimilarityFunction}
\title{NullSimilarityFunction}
\usage{
NullSimilarityFunction(df1, df2, ncombos, MVCorrection, mn)
}
\arguments{
\item{df1}{Original Dataset 1}

\item{df2}{Original Dataset 2}

\item{ncombos}{Null Dataset Samplings}

\item{MVCorrection}{If False, it will ignore missing values rather than count towards the overall average. Default=TRUE}

\item{mn}{The scaling factor. Default will use 1+mean(presence) of a glycopeptide/identification. Setting this to a number will override that process.}
}
\value{
Distribution of Null similarities and their contributions
}
\description{
NullSimilarityFunction
}
