% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservedSimilarityStats.R
\name{ObservedSimilarityStats}
\alias{ObservedSimilarityStats}
\title{ObservedSimilarityStats gets the pertinent summary statistics about the Observed Similarity}
\usage{
ObservedSimilarityStats(TestSimilarity, ActualSim)
}
\arguments{
\item{TestSimilarity}{All Similarity Values in the Test Distribution}

\item{ActualSim}{The Observed Similarity}
}
\value{
A list of the Observed similarity, its percentile location, and its zscore
}
\description{
ObservedSimilarityStats gets the pertinent summary statistics about the Observed Similarity
}
\examples{
#
}
