% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlap.R
\name{Overlap}
\alias{Overlap}
\title{Overlap: DEPRICATED: finds the overlap statistics of two density distributions}
\usage{
Overlap(Dis1, Dis2)
}
\arguments{
\item{Dis1}{Distribution 1}

\item{Dis2}{Distribution 2}
}
\value{
Overlap statistics: Total Percent, False Positive (Alpha), and False Negative (Beta)
}
\description{
Overlap: DEPRICATED: finds the overlap statistics of two density distributions
}
\examples{
#
}
