% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapIdxs.R
\name{OverlapIdxs}
\alias{OverlapIdxs}
\title{OverlapIdxs is used by AlphaIdxs and BetaIdxs to get the indices from the overlap data for the alpha and beta regions respectively}
\usage{
OverlapIdxs(OverlapData, AlpBet)
}
\arguments{
\item{AlpBet}{}
}
\value{

}
\description{
OverlapIdxs is used by AlphaIdxs and BetaIdxs to get the indices from the overlap data for the alpha and beta regions respectively
}
