% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideCollapser.R
\name{PeptideCollapser}
\alias{PeptideCollapser}
\title{PeptideCollapser Subset peptides down to a list of sequons}
\usage{
PeptideCollapser(filename, kmin = 2, rel = TRUE, sequonvector)
}
\arguments{
\item{filename}{dataset to be subset}

\item{kmin}{minimum number of glycopeptide observations Default=2}

\item{rel}{Boolean indicating use of relativization. Default=TRUE}

\item{sequonvector}{Sequon Vector}
}
\value{
Returns dataset broken down by sequon (It does not separate by protein)
}
\description{
PeptideCollapser Subset peptides down to a list of sequons
}
\examples{
#
}
