% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideSegment.R
\name{PeptideSegment}
\alias{PeptideSegment}
\title{PeptideSegment}
\usage{
PeptideSegment(
  filename1,
  filename2,
  kmin = 2,
  simtype = "Tanimoto",
  rel = TRUE,
  OverallName,
  SampleName1,
  SampleName2,
  mn = FALSE,
  logopt = TRUE
)
}
\arguments{
\item{filename1}{dataset 1}

\item{filename2}{dataset 2}

\item{kmin}{minimum observations needed DEFAULT=2}

\item{simtype}{similarity type DEFAULT="Tanimoto"}

\item{rel}{Boolean DEFAULT=TRUE}

\item{OverallName}{Analysis name string}

\item{SampleName1}{Sample name string}

\item{SampleName2}{Sample name string}

\item{mn}{Default=FALSE. In default settings it adjusts the distance scaling by average presence. Setting it to a numeric value will use that as a constant instead. (Recommended 1:2)}

\item{logopt}{Boolean indicating the use of the log transform before relative scaling of abundance. Default=TRUE}
}
\value{
List of glycopeptides by comparison
}
\description{
PeptideSegment
}
\examples{
#
}
