% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresenceCalc_Helper.R
\name{PresenceCalc_Helper}
\alias{PresenceCalc_Helper}
\title{PresenceCalc_Helper}
\usage{
PresenceCalc_Helper(gl, df)
}
\arguments{
\item{gl}{glycopeptide list}

\item{df}{dataset}
}
\value{
The presence of each glycopeptide in a dataset
}
\description{
PresenceCalc_Helper
}
\examples{
#
}
