% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresenceCalc_Helper.R
\name{PresenceCalcBase}
\alias{PresenceCalcBase}
\title{PresenceCalcBase finds the presence of a single vector}
\usage{
PresenceCalcBase(vec)
}
\arguments{
\item{vec}{}
}
\value{

}
\description{
PresenceCalcBase finds the presence of a single vector
}
