% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankingByZScore.R
\name{RankingByZScore}
\alias{RankingByZScore}
\title{RankingByZScore determines rank}
\usage{
RankingByZScore(SimilarityObj)
}
\arguments{
\item{SimilarityObj}{Similarity object}
}
\value{
rankings
}
\description{
RankingByZScore determines rank
}
\examples{
#
}
