% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankingByZScoreV2.R
\name{RankingByZScoreV2}
\alias{RankingByZScoreV2}
\title{RankingByZScoreV2 determines rank}
\usage{
RankingByZScoreV2(SimilarityObj)
}
\arguments{
\item{SimilarityObj}{Similarity object}
}
\value{
rankings
}
\description{
RankingByZScoreV2 determines rank
}
\examples{
#
}
