% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankingByZScoreV3.R
\name{RankingByZScoreV3}
\alias{RankingByZScoreV3}
\title{RankingByZScoreV3 determines rank}
\usage{
RankingByZScoreV3(TestSimObj, NullSimObj)
}
\arguments{
\item{TestSimObj}{Test Similarity object}

\item{NullSimObj}{Null Similarity object}
}
\value{
rankings
}
\description{
RankingByZScoreV3 determines rank
}
\examples{
#
}
