% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankingCriterion.R
\name{RankingCriterion}
\alias{RankingCriterion}
\title{RankingCriterion}
\usage{
RankingCriterion(SimObj, ExclusionList = "Default")
}
\arguments{
\item{SimObj}{similarity object}

\item{ExclusionList}{failures}
}
\value{
Ranking and QC data
}
\description{
RankingCriterion
}
\examples{
#
}
