% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankingQualityFunction.R
\name{RankingQualityFunction}
\alias{RankingQualityFunction}
\title{RankingQualityFunction combines ranking and QC data and produces the final output}
\usage{
RankingQualityFunction(SimilarityObj, InternalSimObj1, InternalSimObj2)
}
\arguments{
\item{SimilarityObj}{Similarity Object}

\item{InternalSimObj1}{Internal Similarity Object 1}

\item{InternalSimObj2}{Internal Similarity Object 2}
}
\value{
Quality and ranking object
}
\description{
RankingQualityFunction combines ranking and QC data and produces the final output
}
\examples{
#
}
