% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankingQualityFunctionV2.R
\name{RankingQualityFunctionV2}
\alias{RankingQualityFunctionV2}
\title{RankingQualityFunctionV2 combines ranking and QC data and produces the final output}
\usage{
RankingQualityFunctionV2(SimilarityObj, InternalSimObj1, InternalSimObj2)
}
\arguments{
\item{SimilarityObj}{Similarity Object}

\item{InternalSimObj1}{Internal Similarity Object 1}

\item{InternalSimObj2}{Internal Similarity Object 2}
}
\value{
Quality and ranking object
}
\description{
RankingQualityFunctionV2 combines ranking and QC data and produces the final output
}
\examples{
#
}
