% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankingQualityFunctionV3.R
\name{RankingQualityFunctionV3}
\alias{RankingQualityFunctionV3}
\title{RankingQualityFunctionV3 combines ranking and QC data and produces the final output}
\usage{
RankingQualityFunctionV3(
  TestSimObj,
  NullSimObj,
  ActualSimObj,
  InternalSimObj1,
  InternalSimObj2
)
}
\arguments{
\item{TestSimObj}{Test Similarity Object}

\item{NullSimObj}{Null Similarity Object}

\item{ActualSimObj}{Observed/Actual Similarity Object}

\item{InternalSimObj1}{Internal Similarity Object 1}

\item{InternalSimObj2}{Internal Similarity Object 2}
}
\value{
Quality and ranking object
}
\description{
RankingQualityFunctionV3 combines ranking and QC data and produces the final output
}
\examples{
#
}
