% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelDensity.R
\name{RelDensity}
\alias{RelDensity}
\title{Converts a similarity to relative density}
\usage{
RelDensity(SimDist)
}
\arguments{
\item{SimDist}{Tanimoto Similarity Vector}
}
\value{
Relative Density object
}
\description{
Converts a similarity to relative density
}
\examples{
#
}
