% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReRankingFromRAMZISMain.R
\name{ReRankingFromRAMZISMain}
\alias{ReRankingFromRAMZISMain}
\title{ReRankingFromRAMZISMain reformats the data object from RAMZISMain and sends it into RankingQualityFunctionV3. It is meant to allow easier reanalysis with changed ranking info}
\usage{
ReRankingFromRAMZISMain(
  SimilarityObject,
  RankingInfo = "Numerator",
  QualityInfo = "WeightedContributions"
)
}
\arguments{
\item{SimilarityObject}{Output from RAMZISMain}

\item{RankingInfo}{Default="Numerator" is the standard ranking info, "WeightedContributions" changes to the weighted contribution as ranking info}

\item{QualityInfo}{Default="WeightedContributions" is the standard quality info, "Numerator" changes to the weighted contribution as quality info}
}
\value{
#RankingSummary
}
\description{
ReRankingFromRAMZISMain reformats the data object from RAMZISMain and sends it into RankingQualityFunctionV3. It is meant to allow easier reanalysis with changed ranking info
}
\examples{
#
}
