% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullSampler_Helper.R
\name{SAMPLER_Helper_Parity}
\alias{SAMPLER_Helper_Parity}
\title{Sampler_Helper_Parity is used by Null Sampler Helper function to ensure that the datasets are represented equally}
\usage{
SAMPLER_Helper_Parity(combo, coln1, coln2, minsample = 2)
}
\arguments{
\item{combo}{samplings}

\item{coln1}{samples in 1}

\item{coln2}{samples in 2}

\item{minsample}{minimum number of samples needed from each}
}
\value{

}
\description{
Sampler_Helper_Parity is used by Null Sampler Helper function to ensure that the datasets are represented equally
}
