% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMPLER_HelperTIDY.R
\name{SAMPLER_HelperTIDY}
\alias{SAMPLER_HelperTIDY}
\title{SAMPLER_HelperTIDY
Generates samples of a single data set and removes duplicates}
\usage{
SAMPLER_HelperTIDY(coln, sampn = 100, filename = NULL, GroupID = "Group")
}
\arguments{
\item{coln}{Number of columns (samples) in a dataset}

\item{sampn}{Number of samplings desired}

\item{filename}{Source of dataset}

\item{GroupID}{String identifying which group this applies to. ie TestA, TestB, NullA, NullB. Defaults to "Group"}
}
\value{
Matrix of indices for samples
}
\description{
SAMPLER_HelperTIDY
Generates samples of a single data set and removes duplicates
}
\examples{
#
}
