% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityCalculation.R
\name{SimilarityCalculation}
\alias{SimilarityCalculation}
\title{SimilarityCalculation calculates the adjusted tanimoto similarity coefficient for two samplings}
\usage{
SimilarityCalculation(
  ABMatrix,
  AATerm,
  BBTerm,
  PresenceMatrix,
  AMatSize,
  BMatSize,
  DistScale
)
}
\arguments{
\item{ABMatrix}{joint data set matrix}

\item{AATerm}{vector for the A denominator terms}

\item{BBTerm}{vector for the B denominator terms}

\item{PresenceMatrix}{presence weights for comparisons of size equal to ABMatrix}

\item{AMatSize}{number of A samplings}

\item{BMatSize}{number of B samples}

\item{DistScale}{distance scaling term}
}
\value{
Similarity coefficients and the composite parts
}
\description{
SimilarityCalculation calculates the adjusted tanimoto similarity coefficient for two samplings
}
\examples{
#
}
