% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityCalculation_Singular.R
\name{SimilarityCalculation_Singular}
\alias{SimilarityCalculation_Singular}
\title{SimilarityCalculation_Singular
For calculating two datasets similarity without sampling. It is used in RAMZIS_Main to calculate the observed similarity}
\usage{
SimilarityCalculation_Singular(df1, df2, mn = FALSE)
}
\arguments{
\item{df1}{The first dataframe to be compared}

\item{df2}{The second dataframe to be compared}

\item{mn}{The scaling factor. Default will use 1+mean(presence) of a glycopeptide/identification. Setting this to a number will override that process.}
}
\value{

}
\description{
SimilarityCalculation_Singular
For calculating two datasets similarity without sampling. It is used in RAMZIS_Main to calculate the observed similarity
}
