% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityCalculationTIDY.R
\name{SimilarityCalculationTIDY}
\alias{SimilarityCalculationTIDY}
\title{SimilarityCalculationTIDY calculates the adjusted tanimoto similarity coefficient for two samplings}
\usage{
SimilarityCalculationTIDY(A_Obj, B_Obj, AMatSize, BMatSize, IDVec, DistScale)
}
\arguments{
\item{AMatSize}{number of A samplings}

\item{BMatSize}{number of B samplings}

\item{DistScale}{distance scaling term}

\item{ABMatrix}{joint data set matrix}

\item{AATerm}{vector for the A denominator terms}

\item{BBTerm}{vector for the B denominator terms}

\item{PresenceMatrix}{presence weights for comparisons of size equal to ABMatrix}
}
\value{
Similarity coefficients and the composite parts
}
\description{
SimilarityCalculationTIDY calculates the adjusted tanimoto similarity coefficient for two samplings
}
\examples{
#
}
