% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPlot.R
\name{SimPlot}
\alias{SimPlot}
\title{SimPlot plots the Test and Null distributions}
\usage{
SimPlot(
  PlotTitle,
  SimilarityObj,
  legendpos = "topleft",
  verbose = F,
  legendval = FALSE,
  xbound = c(0, 1),
  ybound = T,
  logval = FALSE,
  legendcexval = 1.1
)
}
\arguments{
\item{PlotTitle}{Plot Title string}

\item{SimilarityObj}{Test and Null Similarity Object from SymmetricalSimBootstrap()}

\item{legendpos}{legend position}

\item{verbose}{Boolean to send overlap data to console}

\item{legendval}{Boolean to produce legend DEFAULT=TRUE}

\item{xbound}{xbound vector default c(0,1)}

\item{ybound}{ybound vector default goes to max peak height}

\item{logval}{Boolean indicating log transform before relative abundance. DEFAULT=TRUE}

\item{legendcexval}{legend text size}
}
\value{
Makes plot.
}
\description{
SimPlot plots the Test and Null distributions
}
\examples{
#
}
