% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPlotGeneral.R
\name{SimPlotGeneral}
\alias{SimPlotGeneral}
\title{SimPlotGeneral produces a plot of the densities of the similarity of the Test and Null Similarities}
\usage{
SimPlotGeneral(PlotTitle, dfSim, OverlapData, legend.bool = T)
}
\arguments{
\item{legend.bool}{}
}
\value{

}
\description{
SimPlotGeneral produces a plot of the densities of the similarity of the Test and Null Similarities
}
