% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPlotGGBase.R
\name{SimPlotGGBase}
\alias{SimPlotGGBase}
\title{SimPlotGGBase plots similarity distributions. It is primarily for use through other functions}
\usage{
SimPlotGGBase(
  PlotTitle,
  dfSim,
  RefName,
  OverlapData,
  ObservedSimObj,
  Confidence = NULL,
  legend.bool = T,
  themepick = theme_RAMZIS
)
}
\arguments{
\item{PlotTitle}{The title of the plot}

\item{dfSim}{TIDY Formatted Similarity Distributions}

\item{RefName}{Reference Distribution Name. ("Null","Internal1","Internal2","All")}

\item{OverlapData}{Overlap data from OverlapCalculator}

\item{ObservedSimObj}{List produced by ObservedSimilarityStats()}

\item{Confidence}{Internal Confidence Score. Do not touch if doing TestNull Plot. DEFAULT=NULL}

\item{legend.bool}{Boolean indicating whether or not to produce the legend. Default=T}

\item{themepick}{Default=theme_RAMZIS. Theme object from ggplot2 for formatting.}
}
\value{
A similarity plot using ggplot2
}
\description{
SimPlotGGBase plots similarity distributions. It is primarily for use through other functions
}
\examples{
#
}
