% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPlotInternal.R
\name{SimPlotInternal}
\alias{SimPlotInternal}
\title{SimPlotInternal produces a similarity plot of one of the two internal distributions}
\usage{
SimPlotInternal(
  PlotTitle,
  dfSim,
  Dataset,
  OverlapData,
  Confidence = NULL,
  legend.bool = T
)
}
\arguments{
\item{PlotTitle}{The Plot Title}

\item{dfSim}{Similarity Object Produced by RAMZISMain()}

\item{Dataset}{Selects which dataset to produce the internal plot of. Dataset=1 produces internal of the first dataset given. Dataset=2 produces the internal of the second dataset.}

\item{OverlapData}{OverlapData from RAMZISMain() OverlapCalculator()}

\item{Confidence}{From InternalConfidenceScore()}

\item{legend.bool}{Boolean to produce legend or not}
}
\value{

}
\description{
SimPlotInternal produces a similarity plot of one of the two internal distributions
}
