% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization_SubFunction.R
\name{Standardization_SubFunction}
\alias{Standardization_SubFunction}
\title{Standardization_SubFunction scales a dataframe's columns to all be between 0 and 1}
\usage{
Standardization_SubFunction(df, rel = "Within")
}
\arguments{
\item{df}{Dataframe to be standardized}

\item{rel}{Choice of Standardization method. 'Within' scales compared to largest total signal size in df. 'AsIs' does no scaling. NUMERIC scales to a specified number. Default='Within'}
}
\value{
Standardized dataframe
}
\description{
Standardization_SubFunction scales a dataframe's columns to all be between 0 and 1
}
\examples{
#
}
