% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymmetricalSimBootstrap.R
\name{SymmetricalSimBootstrap}
\alias{SymmetricalSimBootstrap}
\title{SymmetricalSimBootstrap compares two datasets of glycopeptide identifications}
\usage{
SymmetricalSimBootstrap(
  filename1,
  filename2,
  kmin = 2,
  rel = TRUE,
  MVCorrection = TRUE,
  mn = FALSE,
  logopt = TRUE,
  normvec = list("None", "None")
)
}
\arguments{
\item{filename1}{first sample dataset}

\item{filename2}{second sample dataset}

\item{kmin}{minimum number of observations Default=2}

\item{rel}{Boolean designating relativity Default=TRUE}

\item{MVCorrection}{Boolean missing value correction, if disabled it turns 0s to NAs. Default=TRUE}

\item{mn}{Default=FALSE. In default settings it adjusts the distance scaling by average presence. Setting it to a numeric value will use that as a constant instead. (Recommended 1:2)}

\item{logopt}{Boolean indicating the use of the log transform before relative scaling of abundance. Default=TRUE}

\item{normvec}{Optional: the normalization vectors to adjust signal abundance between samples. Default=list('None','None')}
}
\value{
Similarity Object
}
\description{
SymmetricalSimBootstrap compares two datasets of glycopeptide identifications
}
\examples{
#
}
