% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymmetricalSimBootstrapv2.R
\name{SymmetricalSimBootstrapV2}
\alias{SymmetricalSimBootstrapV2}
\title{SymmetricalSimBootstrapV2 compares two datasets of glycopeptide identifications. It is a temporary updated version}
\usage{
SymmetricalSimBootstrapV2(
  filename1,
  filename2,
  kmin = 2,
  rel = "Joint",
  MVCorrection = TRUE,
  mn = FALSE,
  logopt = TRUE,
  normvec = list("None", "None"),
  rel_force = FALSE
)
}
\arguments{
\item{filename1}{first sample dataset}

\item{filename2}{second sample dataset}

\item{kmin}{minimum number of observations Default=2}

\item{rel}{Boolean designating standardization method. Default='Joint'}

\item{MVCorrection}{Boolean missing value correction, if disabled it turns 0s to NAs. Default=TRUE}

\item{mn}{Default=FALSE. In default settings it adjusts the distance scaling by average presence. Setting it to a numeric value will use that as a constant instead. (Recommended 1:2)}

\item{logopt}{Boolean indicating the use of the log transform before relative scaling of abundance. Default=TRUE}

\item{normvec}{Optional: the normalization vectors to adjust signal abundance between samples. Default=list('None','None')}

\item{rel_force}{Boolean to force relativization in addition to TIC normalization. Default=FALSE}
}
\value{
Similarity Object
}
\description{
SymmetricalSimBootstrapV2 compares two datasets of glycopeptide identifications. It is a temporary updated version
}
\examples{
#
}
