% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestMembershipProportion.R
\name{TestMembershipProportion}
\alias{TestMembershipProportion}
\title{TestMembershipProportion is for determining the membership of samples to peaks to the Internal similarity distribution}
\usage{
TestMembershipProportion(SimDist, BootDis1, BootDis2, ModalityList)
}
\arguments{
\item{SimDist}{Similarity Distribution}

\item{BootDis1}{Bootstrapping of the first sample}

\item{BootDis2}{Bootstrapping of the second sample}

\item{ModalityList}{Output from Mo}
}
\value{
Sample Proportionship
}
\description{
TestMembershipProportion is for determining the membership of samples to peaks to the Internal similarity distribution
}
\examples{
#
}
