% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheoreticalDataGenerator.R
\name{TheoreticalDataGenerator}
\alias{TheoreticalDataGenerator}
\title{TheoreticalDataGenerator makes a theoretical dataset of relative abundances}
\usage{
TheoreticalDataGenerator(
  n,
  g,
  p,
  w,
  alp = 2,
  bet = 2,
  maxim = TRUE,
  w0 = 2,
  w1 = 10
)
}
\arguments{
\item{n}{number of samples}

\item{g}{number of identifications}

\item{p}{average presence}

\item{w}{controls variance; if given a vector length equal to g the variance is controlled by identification; if given a vector length equal to n the variance is controlled by sample}

\item{alp}{controls mean of the beta distribution used for sampling, if given a vector length g each identification has its own average}

\item{maxim}{Boolean maximizing each column's identification}

\item{w0}{number of significant digits to round alpha and beta to after finding the means}

\item{w1}{controls the variance of the beta distribution production}

\item{betcontrols}{mean of the beta distribution used for sampling, if given a vector length g each identification has its own average}
}
\value{
Theoretical dataset
}
\description{
TheoreticalDataGenerator makes a theoretical dataset of relative abundances
}
\examples{
data<-TheoreticalDataGenerator(5,20,0.7,5.5,alp=seq(100,2000,by=100),bet=seq(2000,100,by=-100))
#data$Dataset is a dataset with 5 samples and 20 identifications (glycopeptides)
#the means of data will trend from 0 to 1
}
