% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyContributions.R
\name{TidyContributions}
\alias{TidyContributions}
\title{TidyContributions reformats the similarity contribution data into TIDY friendly format}
\usage{
TidyContributions(
  SimObject,
  Boot1,
  Boot2,
  GPID = NULL,
  CompID = "X_",
  IntBool = T
)
}
\arguments{
\item{SimObject}{Similarity Object}

\item{Boot1}{Sampling of dataset 1}

\item{Boot2}{Sampling of dataset 2}

\item{GPID}{Glycopeptide IDs. Default=NULL}

\item{CompID}{Comparison ID. Default='X_'}

\item{IntBool}{Boolean indicating if an internal similarity comparison or not. Default assumes internal. Default=T}
}
\value{
TIDY formatted contribution data
}
\description{
TidyContributions reformats the similarity contribution data into TIDY friendly format
}
