% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorMatrixMean.R
\name{VectorMatrixMean}
\alias{VectorMatrixMean}
\title{VectorMatrixMean determines the mean of a dataframe against a single vector}
\usage{
VectorMatrixMean(vector1, dataframe1)
}
\arguments{
\item{vector1}{numeric vector}

\item{dataframe1}{numeric dataframe/matrix}
}
\value{
matrix of means
}
\description{
VectorMatrixMean determines the mean of a dataframe against a single vector
}
\examples{
#
}
