% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorMatrixMultiplication.R
\name{VectorMatrixMultiplication}
\alias{VectorMatrixMultiplication}
\title{VectorMatrixMultiplication determines the multiplication of a dataframe against a single vector}
\usage{
VectorMatrixMultiplication(vector1, dataframe1)
}
\arguments{
\item{vector1}{numeric vector}

\item{dataframe1}{numeric dataframe/matrix}
}
\value{
matrix of multiplications
}
\description{
VectorMatrixMultiplication determines the multiplication of a dataframe against a single vector
}
\examples{
#
}
