% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZScoreContributionFunction.R
\name{ZScoreContributionFunction}
\alias{ZScoreContributionFunction}
\title{ZScoreContributionFunction determines ZScore ranking between test and null}
\usage{
ZScoreContributionFunction(ReferenceCont, TestCont)
}
\arguments{
\item{ReferenceCont}{Null distribution}

\item{TestCont}{Test similarity distribution}
}
\value{
z scores for determining ranks
}
\description{
ZScoreContributionFunction determines ZScore ranking between test and null
}
\examples{
#
}
