% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZScoreContributionFunctionV3.R
\name{ZScoreContributionFunctionV3}
\alias{ZScoreContributionFunctionV3}
\title{ZScoreContributionFunctionV3 determines ZScore ranking between test and null}
\usage{
ZScoreContributionFunctionV3(ReferenceCont, TestCont)
}
\arguments{
\item{ReferenceCont}{Null distribution}

\item{TestCont}{Test similarity distribution}
}
\value{
z scores for determining ranks
}
\description{
ZScoreContributionFunctionV3 determines ZScore ranking between test and null
}
\examples{
#
}
