% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZScoreDistributionFunction.R
\name{ZScoreDistributionFunction}
\alias{ZScoreDistributionFunction}
\title{ZScoreDistributionFunction determines the Z-score of all values in the second distribution compared to the mean and SD of the first}
\usage{
ZScoreDistributionFunction(ReferenceDis, TestDis)
}
\arguments{
\item{ReferenceDis}{First distribution for reference}

\item{TestDis}{Second distribution for testing}
}
\value{
distribution of z scores
}
\description{
ZScoreDistributionFunction determines the Z-score of all values in the second distribution compared to the mean and SD of the first
}
\examples{
#
}
