% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZScoreFunction.R
\name{ZScoreFunction}
\alias{ZScoreFunction}
\title{ZScoreFunction determines the z score.}
\usage{
ZScoreFunction(ReferenceMean, ReferenceSD, TestNum)
}
\arguments{
\item{ReferenceMean}{A mean}

\item{ReferenceSD}{the Standard Deviation}

\item{TestNum}{The Number you want to know the z-score of}
}
\value{
the z-score
}
\description{
ZScoreFunction determines the z score.
}
\examples{
#
}
