% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapIdxs.R
\name{AlphaIdxs}
\alias{AlphaIdxs}
\title{AlphaIdxs finds the indices from the overlap data for the alpha region}
\usage{
AlphaIdxs(OverlapData)
}
\arguments{
\item{OverlapData}{PlaceHolder}

\item{AlpBet}{PlaceHolder}
}
\value{
PlaceHolder
}
\description{
AlphaIdxs finds the indices from the overlap data for the alpha region
}
\examples{
 #
}
