% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlycReRead.R
\name{GlycReRead}
\alias{GlycReRead}
\title{GlycReRead: Read a set of GlycReSoft files and collate them into a joint dataset}
\usage{
GlycReRead(Filelist, Outfilename = NULL, verbose = T)
}
\arguments{
\item{Filelist}{A vector of exported GlycReSoft results files (.csv) of at least length 1.}

\item{Outfilename}{The save location for the collective dataset (don't include a filetype). If unused results will not be saved.}

\item{verbose}{Default: TRUE  Returns dataset to environment}
}
\value{
Matrix of glycopeptide abundances with glycopeptides as rows and samples as columns
}
\description{
GlycReRead: Read a set of GlycReSoft files and collate them into a joint dataset
}
\examples{
#files<-c('gsoft1.csv','gsoft2.csv','gsoft3.csv')
#AbundanceDF<-GlycReRead(files,'OutputSaveFile')
}
