% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN_OUTPUT_GATHERER.R
\name{MAIN_OUTPUT_GATHERER}
\alias{MAIN_OUTPUT_GATHERER}
\title{MAIN_OUTPUT_GATHERER is used by RAMZISMain() to collect and organize all outputs of the similarity comparisons}
\usage{
MAIN_OUTPUT_GATHERER(
  Internal1,
  Internal2,
  Test,
  Null,
  Observed,
  ITO1,
  ITO2,
  NTO,
  Boot1,
  Boot2,
  NullBoot1,
  NullBoot2
)
}
\arguments{
\item{NullBoot2}{}
}
\value{

}
\description{
MAIN_OUTPUT_GATHERER is used by RAMZISMain() to collect and organize all outputs of the similarity comparisons
}
