% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalityTest.R
\name{ModalityTest}
\alias{ModalityTest}
\title{ModalityTest tests for over or under-represented membership in peaks}
\usage{
ModalityTest(BootSet, IntSim1, IntSim2)
}
\arguments{
\item{BootSet}{List of samplings produced for the Test and Internal comparisons}

\item{IntObject1}{Internal Similarity}

\item{IntObject2}{Internal Similarity}
}
\value{
Rejection list of samples based on Modality criterion
}
\description{
ModalityTest tests for over or under-represented membership in peaks
}
\examples{
#
}
