% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapCaclulator.R
\name{OverlapCalculator}
\alias{OverlapCalculator}
\title{Overlap Calculator}
\usage{
OverlapCalculator(ReferenceDis, TestDis)
}
\arguments{
\item{ReferenceDis}{reference distribution, generally higher}

\item{TestDis}{comparison distribution}
}
\value{
Overlap statistics
}
\description{
Overlap Calculator
}
\examples{
#
}
