% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAMZISMain.R
\name{RAMZISMain}
\alias{RAMZISMain}
\title{RAMZISMain is intended to be the hassle free way to run RAMZIS. It should produce all the similarity comparisons necessary for a given set of files}
\usage{
RAMZISMain(
  filename1,
  filename2,
  alpha = 0.05,
  beta = 0.2,
  conf_thresh = 2,
  int_thresh = 0.25,
  observedbounds = c("ZScore", 3),
  kmin = 2,
  kmin_int = 1,
  rel = TRUE,
  MVCorrection = TRUE,
  mn = FALSE,
  verbose = T,
  logopt = FALSE,
  normvec = c("None", "None")
)
}
\arguments{
\item{filename1}{The first file or dataset of the similarity comparison}

\item{filename2}{The second file or dataset of the similarity comparison}

\item{alpha}{Default=0.05}

\item{beta}{Default=0.20}

\item{conf_thresh}{Default=2}

\item{int_thresh}{Default=0.25}

\item{observedbounds}{Default= c("ZScore",3)}

\item{kmin}{Default=2}

\item{kmin_int}{Default=1}

\item{rel}{Default=TRUE}

\item{MVCorrection}{default=TRUE}

\item{mn}{Default=FALSE}

\item{verbose}{default=TRUE}

\item{logopt}{Default=FALSE}

\item{normvec}{Default=c('None','None')}
}
\value{
Automated RAMZIS Output Currently underdevelopment
}
\description{
RAMZISMain is intended to be the hassle free way to run RAMZIS. It should produce all the similarity comparisons necessary for a given set of files
}
\examples{
#
}
