% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankingByZScore.R
\name{RankingByZScore}
\alias{RankingByZScore}
\title{RankingByZScore determines rank}
\usage{
RankingByZScore(SimilarityObj, RankInfo = "RankInfo")
}
\arguments{
\item{SimilarityObj}{Similarity object from RAMZISMain()}

\item{RankInfo}{Which information to use for ranking. Default="RankInfo", Alternative use is "WeightedContributions"}
}
\value{
rankings
}
\description{
RankingByZScore determines rank
}
\examples{
#
}
