% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMPLER_Helper.R
\name{SAMPLER_Helper}
\alias{SAMPLER_Helper}
\title{SAMPLER_Helper
Generates samples of a single data set and removes duplicates}
\usage{
SAMPLER_Helper(coln, sampn = 100, filename = NULL)
}
\arguments{
\item{coln}{Number of columns (samples) in a dataset}

\item{sampn}{Number of samplings desired}

\item{filename}{Source of dataset}
}
\value{
Matrix of indices for samples
}
\description{
SAMPLER_Helper
Generates samples of a single data set and removes duplicates
}
\examples{
#
}
