% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDataClean.R
\name{SimDataClean}
\alias{SimDataClean}
\title{SimDataClean Standardize and normalize an abundance matrix}
\usage{
SimDataClean(
  filename,
  kmin = 2,
  rel = TRUE,
  normvector = "None",
  logoption = TRUE
)
}
\arguments{
\item{filename}{The abundance matrix in question. This can also be the matrix/df already in system}

\item{kmin}{The minimum number of observations needed for a glycopeptide to be real. Default=2}

\item{rel}{Pick if values should be normalized relative to the max total abundance of any column(TRUE) or read as is(FALSE). Default = TRUE}

\item{normvector}{A vector for normalizing signal between samples, usually based on the XIC. Default: 'None' reads values as is; 'Relative' scales samples to their overall signal (assuming equal signal distribution). Providing a vector multiplies sample glycopeptides by that vector}

\item{logoption}{Boolean indicating use of log transformation on values. This occurs after normalization and before relativity. Default= TRUE}
}
\value{
The normalized and standardized relative log abundance dataframe.
}
\description{
SimDataClean Standardize and normalize an abundance matrix
}
\examples{
#From the outputs of GlycReRead example
#Sample1DataMatrix<-SimDataClean('OutputSaveFile.csv')
## OR
#Sample1DataMatric<-SimDataClean(AbundanceDF)
}
