% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPlotGGBase.R
\name{SimPlotGGBase}
\alias{SimPlotGGBase}
\title{SimPlotGGBase plots similarity distributions. It is primarily for use through other functions}
\usage{
SimPlotGGBase(
  PlotTitle,
  dfSim,
  RefName,
  OverlapData,
  Confidence = NULL,
  legend.bool = T
)
}
\arguments{
\item{PlotTitle}{The title of the plot}

\item{dfSim}{PlaceHolder}

\item{RefName}{PlaceHolder}

\item{OverlapData}{PlaceHolder}

\item{Confidence}{PlaceHolder}

\item{legend.bool}{PlaceHolder}
}
\value{
PlaceHolder
}
\description{
SimPlotGGBase plots similarity distributions. It is primarily for use through other functions
}
\examples{
#
}
