% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPlotInternal.R
\name{SimPlotInternal}
\alias{SimPlotInternal}
\title{SimPlotInternal produces a similarity plot of one of the two internal distributions}
\usage{
SimPlotInternal(
  PlotTitle,
  SimilarityObject,
  Dataset,
  legend.bool = T,
  themepick = theme_RAMZIS,
  zopt = T,
  ...
)
}
\arguments{
\item{PlotTitle}{The Plot Title}

\item{SimilarityObject}{Similarity Object Produced by RAMZISMain()}

\item{Dataset}{Selects which dataset to produce the internal plot of. Dataset=1 produces internal of the first dataset given. Dataset=2 produces the internal of the second dataset.}

\item{legend.bool}{Boolean to produce legend or not}

\item{themepick}{Default=theme_RAMZIS. Theme object from ggplot2 for formatting.}

\item{zopt}{Indicates use of Z-score for simulation validity over percentile. Default=T}
}
\value{
Internal Similarity Plot
}
\description{
SimPlotInternal produces a similarity plot of one of the two internal distributions
}
\examples{
#
}
