% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheoreticalDataGenerator.R
\name{TheoreticalDataGenerator}
\alias{TheoreticalDataGenerator}
\title{TheoreticalDataGenerator makes a theoretical dataset of relative abundances}
\usage{
TheoreticalDataGenerator(
  n,
  g,
  p,
  samplevar = 1,
  alp = 2,
  bet = 2,
  maxim = FALSE,
  ratio_sig_N = 2,
  rat_mult = 10,
  fakeGPLeader = "Identification"
)
}
\arguments{
\item{n}{number of samples}

\item{g}{number of identifications}

\item{p}{average presence}

\item{samplevar}{controls variance; lower is less variant if given a vector length equal to g the variance is controlled by identification; if given a vector length equal to n the variance is controlled by sample. default=0}

\item{alp}{controls mean of the beta distribution used for sampling, if given a vector length g each identification has its own average}

\item{maxim}{Boolean maximizing each column's identification. Default=FALSE}

\item{ratio_sig_N}{number of significant digits to round alpha and beta to after finding the means. requires integers. default=2}

\item{rat_mult}{controls the variance of the beta distribution production. default=10}

\item{fakeGPLeader}{name of identifications produced}

\item{betcontrols}{mean of the beta distribution used for sampling, if given a vector length g each identification has its own average}
}
\value{
Theoretical dataset
}
\description{
TheoreticalDataGenerator makes a theoretical dataset of relative abundances
}
\examples{
#data<-TheoreticalDataGenerator(5,20,0.7,5.5,alp=seq(100,2000,by=100),bet=seq(2000,100,by=-100))
#data$Dataset is a dataset with 5 samples and 20 identifications (glycopeptides)
#the means of data will trend from 0 to 1
}
