% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorMatrixAddition.R
\name{VectorMatrixAddition}
\alias{VectorMatrixAddition}
\title{VectorMatrixAddition determines the multiplication of a dataframe against a single vector}
\usage{
VectorMatrixAddition(vector1, dataframe1)
}
\arguments{
\item{vector1}{numeric vector}

\item{dataframe1}{numeric dataframe/matrix}
}
\value{
matrix of additions
}
\description{
VectorMatrixAddition determines the multiplication of a dataframe against a single vector
}
\examples{
#
}
